+++ src/HexagoScreenSaverConfig.hpp
@@ -37,7 +37,7 @@ namespace hexago {
          * should be covered by Hexagons at any point in time. Also given as a
          * percentage where 1.0 is 100% and 0.0 is 0%
          */
-        float minimum_screen_cover;
+        double minimum_screen_cover;
         // spawn mode specifier, refer to enum description
         hexagon_spawn_mode_t spawn_mode;
         // background colour mode specifier, refer to enum description
@@ -45,15 +45,15 @@ namespace hexago {
 
         // main constructor
         HexagoScreenSaverConfig(
-            ParameterRange<float> start_size_range,
-            ParameterRange<float> decay_speed_range,
+            ParameterRange<double> start_size_range,
+            ParameterRange<double> decay_speed_range,
             colour_model_t colour_model,
             ParameterRange<colour_channel_t> d_colour_channel_range,
             ParameterRange<colour_channel_t> e_colour_channel_range,
             ParameterRange<colour_channel_t> f_colour_channel_range,
             ParameterRange<colour_channel_t> alpha_colour_channel_range,
             unsigned int framerate,
-            float minimum_screen_cover,
+            double minimum_screen_cover,
             hexagon_spawn_mode_t spawn_mode,
             background_colour_t background_colour
         );
@@ -62,7 +62,7 @@ namespace hexago {
         HexagoScreenSaverConfig(
             HexagonFactoryConfig sub_config,
             unsigned int framerate,
-            float minimum_screen_cover,
+            double minimum_screen_cover,
             hexagon_spawn_mode_t spawn_mode,
             background_colour_t background_colour
         );
